Extract common yaml utils.
