+++ types/certificate_generator.go
@@ -11,7 +11,6 @@ import (
 	"time"
 
 	"github.com/cloudfoundry/bosh-utils/errors"
-	"gopkg.in/yaml.v2"
 )
 
 type CertificateGenerator struct {
@@ -183,39 +182,4 @@ func generateCertResponse(privateKey *rsa.PrivateKey, certificateRaw, rootCARaw 
 	}
 
 	return certResponse
-}
-
-func stringInArray(key string, list []string) bool {
-	for _, value := range list {
-		if key == value {
-			return true
-		}
-	}
-	return false
-}
-
-func objToStruct(input interface{}, str interface{}, supportedParameters []string) error {
-	valBytes, err := yaml.Marshal(input)
-	if err != nil {
-		return errors.WrapErrorf(err, "Expected input to be serializable")
-	}
-
-	parametersMap := make(map[string]interface{})
-	err = yaml.Unmarshal(valBytes, parametersMap)
-	if err != nil {
-		return errors.WrapErrorf(err, "Expected input to be deserializable")
-	}
-
-	for key := range parametersMap {
-		if !stringInArray(key, supportedParameters) {
-			return errors.Errorf("Unsupported parameter '%s'", key)
-		}
-	}
-
-	err = yaml.Unmarshal(valBytes, str)
-	if err != nil {
-		return errors.WrapErrorf(err, "Expected input to be deserializable")
-	}
-
-	return nil
 }
