+++ types/password_generator_test.go
@@ -30,6 +30,13 @@ var _ = Describe("PasswordGenerator", func() {
 				Expect(len(password.(string))).To(Equal(32))
 			})
 
+			It("errors on unknown parameters", func() {
+				params := map[interface{}]interface{}{"unsupported": 32}
+				_, err := generator.Generate(params)
+				Expect(err.Error()).ToNot(BeNil())
+				Expect(err.Error()).To(Equal("Failed to generate password, parameters are invalid: Unsupported parameter 'unsupported'"))
+			})
+
 			It("generates unique passwords", func() {
 				password1, err := generator.Generate(nil)
 				Expect(err).ToNot(HaveOccurred())
